+++ db/eagle/src/models/offer/SmartOffer.js
@@ -1,6 +1,6 @@
 const Mongoose = require("mongoose");
 const objectIdType = Mongoose.Schema.Types.ObjectId;
-const DoubleType = Mongoose.Schema.Types.Double;
+const DoubleType = Mongoose.Schema.Types.Decimal128;
 
 const { config } = require("../../constants/index");
 const { getConstLabel, getConstValue } = require("../../helper/Util");
